requirements.txt
flask==3.0.0
pyyaml==6.0.1
jinja2==3.1.2
weasyprint==61.0
